requirements.txt
# psycopg2==2.9.2
Django==3.2.9
djangorestframework==3.13.1
Pillow==8.4.0
Markdown==3.3.6
django-filter==21.1
psycopg2-binary==2.9.3
# Pillow >= /.0
django-cors-headers==3.10.1
django-admin-row-actions==0.0.5